+++ packages/reactivity/src/signal.js
@@ -9,20 +9,20 @@ import {
   unique,
   wrapFunction,
 } from '@semantic-ui/utils';
+
 import { Dependency } from './dependency.js';
 import { Reaction } from './reaction.js';
+
 export class Signal {
   constructor(initialValue, { context, equalityFunction, allowClone = true, cloneFunction } = {}) {
-
     // pass in some metadata for debugging
     this.dependency = new Dependency({
       firstRun: true,
-      value: initialValue
+      value: initialValue,
     });
 
     // allow user to opt out of value cloning
     this.allowClone = allowClone;
-
 
     // allow custom equality function
     this.equalityFunction = (equalityFunction)
@@ -40,22 +40,23 @@ export class Signal {
     this.setContext(context);
   }
 
-  // set debugging context for signal
+  // set debugging context for signal removing any present context
   setContext(additionalContext = {}) {
     const defaultContext = {
       value: this.currentValue,
     };
     this.context = {
       ...defaultContext,
-      ...additionalContext
+      ...additionalContext,
     };
   }
 
+  // add context to signal
   addContext(additionalContext = {}) {
-    if(!this.context) {
+    if (!this.context) {
       this.context = {};
     }
-    for(const key in additionalContext) {
+    for (const key in additionalContext) {
       this.context[key] = additionalContext[key];
     }
   }
@@ -101,7 +102,7 @@ export class Signal {
   set value(newValue) {
     if (!this.equalityFunction(this.currentValue, newValue)) {
       this.currentValue = this.maybeClone(newValue);
-      this.setContext();
+      this.addContext({ value: newValue });
       this.setTrace();
       this.dependency.changed(this.context);
     }
@@ -113,7 +114,6 @@ export class Signal {
 
   set(newValue) {
     if (!this.equalityFunction(this.currentValue, newValue)) {
-      this.addContext({ value: newValue });
       this.value = newValue;
     }
   }
@@ -132,43 +132,62 @@ export class Signal {
     return this.set(undefined);
   }
 
+  // mutate the current value by a mutation function
+  mutate(mutationFn) {
+    // we use clone in all cases to detect for changes only
+    const beforeClone = this.clone(this.currentValue);
+    const result = mutationFn(this.currentValue);
+
+    if (result !== undefined) {
+      // if the mutation returned a value just set it
+      this.value = result;
+    }
+    else {
+      // if no value returned check if the value changed from side effects
+      // in this case we want to trigger reactivity
+      if (!this.equalityFunction(beforeClone, this.currentValue)) {
+        this.setContext();
+        this.setTrace();
+        this.dependency.changed(this.context);
+      }
+    }
+  }
+
   // array helpers
   push(...args) {
-    const arr = this.peek();
-    arr.push(...args);
-    this.set(arr);
+    return this.mutate(arr => {
+      arr.push(...args);
+    });
   }
   unshift(...args) {
-    const arr = this.peek();
-    arr.unshift(...args);
-    this.set(arr);
+    return this.mutate(arr => {
+      arr.unshift(...args);
+    });
   }
   splice(...args) {
-    const arr = this.peek();
-    arr.splice(...args);
-    this.set(arr);
+    return this.mutate(arr => {
+      arr.splice(...args);
+    });
   }
   map(mapFunction) {
-    const newValue = Array.prototype.map.call(this.peek(), mapFunction);
-    this.set(newValue);
+    return this.mutate(arr => Array.prototype.map.call(arr, mapFunction));
   }
   filter(filterFunction) {
-    const newValue = Array.prototype.filter.call(this.peek(), filterFunction);
-    this.set(newValue);
+    return this.mutate(arr => Array.prototype.filter.call(arr, filterFunction));
   }
 
   getIndex(index) {
     return this.get()[index];
   }
   setIndex(index, value) {
-    let arr = this.peek();
-    arr[index] = value;
-    this.set(arr);
+    return this.mutate(arr => {
+      arr[index] = value;
+    });
   }
   removeIndex(index) {
-    let arr = this.peek();
-    arr.splice(index, 1);
-    this.set(arr);
+    return this.mutate(arr => {
+      arr.splice(index, 1);
+    });
   }
 
   // sets
@@ -182,6 +201,9 @@ export class Signal {
       value = property;
       property = indexOrProperty;
     }
+    if(index === -1) {
+      return;
+    }
     const newValue = this.peek().map((object, currentIndex) => {
       if (index == 'all' || currentIndex == index) {
         object[property] = value;
@@ -192,26 +214,30 @@ export class Signal {
   }
 
   toggle() {
-    return this.set(!this.peek());
+    return this.mutate(val => !val);
   }
 
   increment(amount = 1, max) {
-    let newAmount = this.peek() + amount;
-    if(isNumber(max) && newAmount > max) {
-      newAmount = max;
-    }
-    return this.set(newAmount);
+    return this.mutate(val => {
+      let newAmount = val + amount;
+      if (isNumber(max) && newAmount > max) {
+        newAmount = max;
+      }
+      return newAmount;
+    });
   }
   decrement(amount = 1, min) {
-    let newAmount = this.peek() - amount;
-    if(isNumber(min) && newAmount < min) {
-      newAmount = min;
-    }
-    return this.set(newAmount);
+    return this.mutate(val => {
+      let newAmount = val - amount;
+      if (isNumber(min) && newAmount < min) {
+        newAmount = min;
+      }
+      return newAmount;
+    });
   }
 
   now() {
-    return this.set(new Date());
+    return this.mutate(() => new Date());
   }
 
   getIDs(item) {
@@ -227,12 +253,18 @@ export class Signal {
     return this.getID(item) === id;
   }
   getItem(id) {
+    const index = this.getItemIndex(id);
+    if(index !== -1) {
+      return this.getIndex(index);
+    }
+  }
+  getItemIndex(id) {
     return findIndex(this.currentValue, item => this.hasID(item, id));
   }
   setProperty(idOrProperty, property, value) {
     if (arguments.length == 3) {
       const id = idOrProperty;
-      const index = this.getItem(id);
+      const index = this.getItemIndex(id);
       return this.setArrayProperty(index, property, value);
     }
     else {
@@ -244,9 +276,9 @@ export class Signal {
     }
   }
   replaceItem(id, item) {
-    return this.setIndex(this.getItem(id), item);
+    return this.setIndex(this.getItemIndex(id), item);
   }
   removeItem(id) {
-    return this.removeIndex(this.getItem(id));
+    return this.removeIndex(this.getItemIndex(id));
   }
 }